test repo volume loading into container; automate testing of file cp or sharing with container; make it easy to save test files after running tests

+++ tests/test_core.py
@@ -43,7 +43,7 @@ class DockerUtils(object):
         """ get contents of a file in a container
 
         Args:
-            container (:obj:`type of arg_1`): a Docker container
+            container (:obj:`docker.models.containers.container`): a Docker container
             file (:obj:`str`): path to a file in `container`
 
         Returns:
@@ -58,23 +58,62 @@ class DockerUtils(object):
         else:
             raise wc_env.EnvError("cat {} fails".format(file))
 
+    @staticmethod
+    def cmp_files(testcase, container, container_filename, host_file_content=None, host_filename=None):
+        """ Make testcase comparison of content of file on host with file in container
+
+        At least one of `host_file_content` or `host_filename` must be provided.
+        Will not scale to huge files.
+
+        Args:
+            testcase (:obj:`testcase.TestCase`): testcase being run
+            container (:obj:`docker.models.containers.container`): Docker container storing `container_filename`
+            container_filename (:obj:`str`): pathname of file in `container`
+            host_file_content (:obj:`str`, optional): content of host file being compared
+            host_filename (:obj:`str`, optional): pathname of host file being compared
+
+        Raises:
+            :obj:`ValueError`: if both `host_file_content` and `host_filename` are `None`
+        """
+        if host_file_content is None and host_filename is None:
+            raise ValueError('either host_file_content or host_filename must be provided')
+        if host_file_content is None:
+            host_filename = os.path.abspath(os.path.expanduser(host_filename))
+            with open(host_filename, 'r') as f:
+                host_file_content = ''.join(f.readlines())
+        testcase.assertEqual(DockerUtils.get_file(container, container_filename), host_file_content)
+
 
 # todo: port to and test on Windows
 class TestManageContainer(unittest.TestCase):
 
     def setUp(self):
-        self.temp_dir  = tempfile.TemporaryDirectory()
-        self.test_dir = self.temp_dir.name
+        # put in temp dir in /private/tmp which can contain a Docker volume by default
+        # todo: make this OS portable
+        # use mkdtemp() instead of TemporaryDirectory() so files can survive testing for debugging container
+        self.test_dir = tempfile.mkdtemp(dir='/private/tmp')
         self.temp_dir_in_home  = \
-            tempfile.TemporaryDirectory(dir=os.path.abspath(os.path.expanduser('~/tmp')))
-        self.test_dir_in_home = os.path.join('~/tmp', os.path.basename(self.temp_dir_in_home.name))
+            tempfile.mkdtemp(dir=os.path.abspath(os.path.expanduser('~/tmp')))
+        self.test_dir_in_home = os.path.join('~/tmp', os.path.basename(self.temp_dir_in_home))
         self.relative_path_file = os.path.join('tests', 'fixtures', 'relative_path_file.txt')
         self.absolute_path_file = os.path.join(os.getcwd(), self.relative_path_file)
         self.relative_temp_path = os.path.join('tests', 'tmp')
         self.moving_text = 'I Have a Dream'
         with open(self.absolute_path_file, 'w') as f:
             f.write(self.moving_text)
+
         self.sample_repo = os.path.join(os.path.dirname(__file__), 'fixtures', 'sample_repo')
+        self.test_wc_repos = [
+            # test path in home dir
+            os.path.join(self.test_dir_in_home, 'repo_a'),
+            # test full pathname
+            os.path.join(self.test_dir, 'repo_b'),
+            # test relative pathname
+            os.path.join(self.relative_temp_path, 'repo_c')
+        ]
+        for test_wc_repo in self.test_wc_repos:
+            self.make_test_repo(test_wc_repo)
+
         self.docker_client = docker.from_env()
         self.tmp_containers = []
 
@@ -85,42 +124,38 @@ class TestManageContainer(unittest.TestCase):
                 container.remove(force=True)
             except docker.errors.APIError:
                 pass
-        # remove relative temp files
-        shutil.rmtree(self.relative_temp_path, ignore_errors=True)
+        # remove temp files
+        remove_temp_files = True
+        if remove_temp_files:
+            shutil.rmtree(self.relative_temp_path, ignore_errors=True)
+            shutil.rmtree(self.test_dir)
+            shutil.rmtree(self.temp_dir_in_home)
 
-    '''
     @classmethod
     def tearDownClass(cls):
-        DockerUtils.list_all()
-    '''
+        # DockerUtils.list_all()
+        pass
 
     def make_test_repo(self, relative_path):
         # copy contents of self.sample_repo to a test repo at `relative_path`
         test_repo = os.path.abspath(os.path.expanduser(relative_path))
         shutil.copytree(self.sample_repo, test_repo)
+        # create a unique repo name file
+        with open(os.path.join(test_repo, 'REPO_NAME'), 'w') as f:
+            f.write(os.path.basename(test_repo))
 
     def test_constructor(self):
-        test_wc_repos = [
-            # test path in home dir
-            os.path.join(self.test_dir_in_home, 'repo_a'),
-            # test full pathname
-            os.path.join(self.test_dir, 'repo_b'),
-            # test relative pathname
-            os.path.join(self.relative_temp_path, 'repo_c')
-        ]
-        for test_wc_repo in test_wc_repos:
-            self.make_test_repo(test_wc_repo)
-        manage_container = wc_env.ManageContainer(test_wc_repos, '0.1')
-        expected_paths = [
-            os.path.join(self.temp_dir_in_home.name, 'repo_a'),
+        manage_container = wc_env.ManageContainer(self.test_wc_repos, '0.1')
+        expected_repo_paths = [
+            os.path.join(self.temp_dir_in_home, 'repo_a'),
             os.path.join(self.test_dir, 'repo_b'),
             os.path.join(os.getcwd(), self.relative_temp_path, 'repo_c')
         ]
-        for computed_path,expected_path in zip(manage_container.local_wc_repos, expected_paths):
+        for computed_path,expected_path in zip(manage_container.local_wc_repos, expected_repo_paths):
             self.assertEqual(computed_path, expected_path)
         with self.assertRaises(wc_env.EnvError):
             wc_env.ManageContainer([self.absolute_path_file], '0.1')
-        repo_a = os.path.join(self.temp_dir_in_home.name, 'repo_a')
+        repo_a = os.path.join(self.temp_dir_in_home, 'repo_a')
         os.chmod(repo_a, 0)
         with self.assertRaises(wc_env.EnvError):
             wc_env.ManageContainer([repo_a], '0.1')
@@ -154,18 +189,25 @@ class TestManageContainer(unittest.TestCase):
             manage_container = wc_env.ManageContainer([], '0.1',
                 configs_repo_pwd_file=test_no_such_file, ssh_key=test_no_such_file)
 
-    def test_create(self):
+    def test_create_volume_sharing(self):
         # test volume sharing
-        '''
-        manage_container = wc_env.ManageContainer([], '0.0.1',
-            ssh_key=test_ssh_key,
-            git_config_file=test_git_config_file)
+        manage_container = wc_env.ManageContainer(self.test_wc_repos, '0.0.1')
         manage_container.create()
-        self.tmp_containers.append(manage_container.container)
+        container = manage_container.container
+        self.tmp_containers.append(container)
+        # print('docker attach', container.name)
+        # spot-check files in the 3 repos in the container
+        for local_wc_repo in manage_container.local_wc_repos:
+            container_wc_repo_dir = os.path.join(manage_container.container_repo_dir,
+                os.path.basename(local_wc_repo))
+            container_core_py_file = os.path.join(container_wc_repo_dir, 'wc_env/core.py')
+            DockerUtils.cmp_files(self, container, container_core_py_file,
+                host_filename=os.path.join(self.sample_repo, 'wc_env/core.py'))
+            container_REPO_NAME_file = os.path.join(container_wc_repo_dir, 'REPO_NAME')
+            DockerUtils.cmp_files(self, container, container_REPO_NAME_file,
+                host_file_content=os.path.basename(local_wc_repo))
 
-        self.assertEqual(manage_container.container.status, 'created')
-        '''
-
+    def test_create_file_copying(self):
         # test file copying
         test_ssh_key = os.path.join(self.test_dir_in_home, 'test_ssh_key')
         test_ssh_key_content = 'Four score ...\nago our ...'
@@ -181,12 +223,11 @@ class TestManageContainer(unittest.TestCase):
         manage_container.create()
         self.tmp_containers.append(manage_container.container)
 
-        self.assertEqual(DockerUtils.get_file(manage_container.container, '/root/.ssh/id_rsa'),
-            test_ssh_key_content)
-        self.assertEqual(DockerUtils.get_file(manage_container.container, '/root/.ssh/id_rsa.pub'),
-            test_ssh_key_content+'.pub')
-        self.assertEqual(DockerUtils.get_file(manage_container.container, '/root/.gitconfig'),
-            ''.join(open(test_git_config_file, 'r').readlines()))
+        DockerUtils.cmp_files(self, manage_container.container, '/root/.ssh/id_rsa', host_filename=test_ssh_key)
+        DockerUtils.cmp_files(self, manage_container.container, '/root/.ssh/id_rsa.pub',
+            host_file_content=test_ssh_key_content+'.pub')
+        DockerUtils.cmp_files(self, manage_container.container, '/root/.gitconfig',
+            host_filename=test_git_config_file)
 
     def create_test_container(self):
         manage_container = wc_env.ManageContainer([], '0.0.1')
@@ -206,5 +247,4 @@ class TestManageContainer(unittest.TestCase):
             manage_container.cp(self.absolute_path_file, '/root/tmp/no such dir/')
         path_in_container = os.path.join('/tmp/', os.path.basename(self.absolute_path_file))
         manage_container.cp(self.absolute_path_file, path_in_container)
-        self.assertEqual(DockerUtils.get_file(manage_container.container, path_in_container),
-            self.moving_text)
+        DockerUtils.cmp_files(self, manage_container.container, path_in_container, host_file_content=self.moving_text)
